Change `ShowDateValidationError` effect as view effect
