+++ app/src/main/java/org/simple/clinic/bloodsugar/entry/BloodSugarEntryEffectHandler.kt
@@ -62,7 +62,6 @@ class BloodSugarEntryEffectHandler @AssistedInject constructor(
         .subtypeEffectHandler<BloodSugarEntryEffect, BloodSugarEntryEvent>()
         .addConsumer(ShowBloodSugarEntryScreen::class.java, { showBloodSugarEntryScreen(it.date) }, schedulersProvider.ui())
         .addTransformer(PrefillDate::class.java, prefillDate(schedulersProvider.ui()))
-        .addConsumer(ShowDateValidationError::class.java, { showDateValidationError(it.result) }, schedulersProvider.ui())
         .addTransformer(CreateNewBloodSugarEntry::class.java, createNewBloodSugarEntryTransformer())
         .addTransformer(FetchBloodSugarMeasurement::class.java, fetchBloodSugarMeasurement(schedulersProvider.io()))
         .addTransformer(UpdateBloodSugarEntry::class.java, updateBloodSugarEntryTransformer(schedulersProvider.io()))
@@ -115,14 +114,6 @@ class BloodSugarEntryEffectHandler @AssistedInject constructor(
   private fun convertToLocalDate(prefillDate: PrefillDate): LocalDate {
     val instant = if (prefillDate is PrefillSpecificDate) prefillDate.date else Instant.now(userClock)
     return instant.toLocalDateAtZone(userClock.zone)
-  }
-
-  private fun showDateValidationError(result: UserInputDateValidator.Result) {
-    when (result) {
-      InvalidPattern -> ui.showInvalidDateError()
-      DateIsInFuture -> ui.showDateIsInFutureError()
-      is Valid -> throw IllegalStateException("Date validation error cannot be $result")
-    }.exhaustive()
   }
 
   private fun createNewBloodSugarEntryTransformer(): ObservableTransformer<CreateNewBloodSugarEntry, BloodSugarEntryEvent> {
