+++ app/src/main/java/org/simple/clinic/bloodsugar/entry/BloodSugarEntryUiActions.kt
@@ -22,4 +22,6 @@ interface BloodSugarEntryUiActions {
       measurementType: BloodSugarMeasurementType,
       unitPreference: BloodSugarUnitPreference
   )
+  fun showInvalidDateError()
+  fun showDateIsInFutureError()
 }
